separate app into its own component to make it more clean
